feat: Update MatchAnswer return type to int in QuizRepository
The code changes in this commit update the return type of the `MatchAnswer` method in the `QuizRepository` interface from `int64` to `int`. This change aligns the return type with the actual data being returned and improves the consistency of the codebase.

+++ internal/domain/repositories/quiz_repository.go
@@ -18,7 +18,7 @@ type QuizRepository interface {
 	GetQuizBySubjectID(subjectID string) (*entities.Quiz, error)
 	GetQuizByTeacherID(teacherID string) ([]entities.Quiz, error)
 	CountQuestion(quizID string) (int64, error)
-	MatchAnswer(quizID string, answer []string) (int64, error)
+	MatchAnswer(quizID string, answer []string) (int, error)
 	GetQuizType(quizID string) (string, error)
 	Update(quizID string, quiz *entities.Quiz) error
 	Delete(quizID string) error
